Add costing, ownership buckets, roles/permissions, notifications, and reporting
Implements a large enhancement covering four phases of work:

Phase 1 (data model backbone):
- Weighted-average purchase costing (Inventory.avg_cost), snapshotted onto
  SalesDetail.unit_cost and DamagedStock.unit_cost/loss_amount so historical
  reports don't shift when cost changes later
- Inventory ownership type (TKU Product/Consignment/Titip Jual) moves from
  the Inventory page to Receiving time; Inventory re-keyed by
  (product, warehouse, inventory_type) via a partial unique index
- All quantity fields converted Float to Integer across models/schemas/UI
- Stock Opname gains an Incomplete unit category, routed to Damaged Stock
  on approval same as damaged units

Phase 2 (workflow UX):
- Stock Opname approve/reject gated to Manager/Admin, with an Incomplete
  Qty UI
- Sales Return sale picker gains search; Supplier Return manual entry uses
  searchable name pickers instead of raw IDs
- Stock Movement simplified to warehouse-to-warehouse Transfer only
- Receiving bulk upload (CSV+XLSX) added; CSV+XLSX templates for every
  bulk upload type

Phase 3 (roles, permissions, notifications):
- Permission catalog + per-user override tables layered on top of the
  existing User.role tier; every module route now enforces a permission
  server-side (master-data list endpoints stay open as cross-module
  lookups; writes are gated)
- Role Management screen for granting/revoking permissions per user
- On-the-fly notification badges (pending approvals, low stock, rejected
  receivings awaiting a supplier return) computed from existing data

Phase 4 (reporting):
- Inventory Report and Sales Report with Category/Warehouse/Type filters
  applied at the query level, Excel (.xlsx) export, and per-unit Margin/
  Margin % calculated against vendor purchase cost
- Indexes added to back the new report filters

+++ backend/app/models/sales.py
@@ -16,7 +16,7 @@ class Sales(Base, AuditMixin):
     # ── Legacy per-item columns (NULL for new multi-item sales; populated for
     #    old single-item records and maintained by migration) ──────────────────
     product_id    = Column(Integer, ForeignKey("products.product_id"), nullable=True)
-    quantity      = Column(Float, default=0, nullable=True)
+    quantity      = Column(Integer, default=0, nullable=True)
     unit          = Column(String(20), default="PCS", nullable=True)
     sale_price    = Column(Float, default=0, nullable=True)    # VAT-inclusive unit price
     discount_pct  = Column(Float, default=0, nullable=True)    # 0–100
